LPD-20628 Move WorkflowPortletTab service tracker to a component so it can be used in other places
