+++ modules/apps/portal-workflow/portal-workflow-api/src/main/java/com/liferay/portal/workflow/portlet/tab/WorkflowPortletTabServiceTracker.java
@@ -1,0 +1,18 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2024 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+package com.liferay.portal.workflow.portlet.tab;
+
+/**
+ * @author Feliphe Marinho
+ */
+public interface WorkflowPortletTabServiceTracker {
+
+	public boolean contains(String portalWorkflowTabsName);
+
+	public WorkflowPortletTab getWorkflowPortletTab(
+		String portalWorkflowTabsName);
+
+}
